refactor: remove legacy forms
