--- src/components/forms/ProductForm.tsx
@@ -1,243 +1,0 @@
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Textarea } from "@/components/ui/textarea";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import { Trash2, Edit } from "lucide-react";
-import { useProductsWithCategories, useCreateProduct, useUpdateProduct, useDeleteProduct } from "@/hooks/useProducts";
-import { useCategories } from "@/hooks/useCategories";
-import { ProductWithCategory, ProductFormData } from "@/types/products";
-import { formatarMoeda } from "@/utils/pricing";
-
-export const ProductForm = () => {
-  const [formData, setFormData] = useState<ProductFormData>({
-    name: "",
-    description: "",
-    sku: "",
-    category_id: "",
-    cost_unit: 0,
-    packaging_cost: 0,
-    tax_rate: 0
-  });
-  const [editingId, setEditingId] = useState<string | null>(null);
-
-  const { data: categories = [] } = useCategories();
-  const { data: products = [], isLoading } = useProductsWithCategories();
-  const createMutation = useCreateProduct();
-  const updateMutation = useUpdateProduct();
-  const deleteMutation = useDeleteProduct();
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    // Converter "none" para undefined/null antes de enviar
-    const dataToSubmit = {
-      ...formData,
-      category_id: formData.category_id === "none" ? undefined : formData.category_id,
-      sku: formData.sku || undefined,
-      description: formData.description || undefined
-    };
-    
-    if (editingId) {
-      updateMutation.mutate({ id: editingId, data: dataToSubmit }, {
-        onSuccess: () => {
-          resetForm();
-        }
-      });
-    } else {
-      createMutation.mutate(dataToSubmit, {
-        onSuccess: () => {
-          resetForm();
-        }
-      });
-    }
-  };
-
-  const resetForm = () => {
-    setFormData({
-      name: "",
-      description: "",
-      sku: "",
-      category_id: "",
-      cost_unit: 0,
-      packaging_cost: 0,
-      tax_rate: 0
-    });
-    setEditingId(null);
-  };
-
-  const handleEdit = (product: ProductWithCategory) => {
-    setFormData({
-      name: product.name,
-      description: product.description || "",
-      sku: product.sku || "",
-      category_id: product.category_id || "none",
-      cost_unit: product.cost_unit || 0,
-      packaging_cost: product.packaging_cost || 0,
-      tax_rate: product.tax_rate || 0
-    });
-    setEditingId(product.id);
-  };
-
-  return (
-    <div className="space-y-6">
-      <Card>
-        <CardHeader>
-          <CardTitle>{editingId ? "Editar Produto" : "Novo Produto"}</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-              <div>
-                <Label htmlFor="name">Nome *</Label>
-                <Input
-                  id="name"
-                  value={formData.name}
-                  onChange={(e) => setFormData(prev => ({ ...prev, name: e.target.value }))}
-                  required
-                />
-              </div>
-              
-              <div>
-                <Label htmlFor="sku">SKU</Label>
-                <Input
-                  id="sku"
-                  value={formData.sku}
-                  onChange={(e) => setFormData(prev => ({ ...prev, sku: e.target.value }))}
-                />
-              </div>
-            </div>
-            
-            <div>
-              <Label htmlFor="description">Descrição</Label>
-              <Textarea
-                id="description"
-                value={formData.description}
-                onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-              />
-            </div>
-            
-            <div>
-              <Label htmlFor="category">Categoria</Label>
-              <Select value={formData.category_id} onValueChange={(value) => setFormData(prev => ({ ...prev, category_id: value }))}>
-                <SelectTrigger>
-                  <SelectValue placeholder="Selecione uma categoria" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="none">Nenhuma categoria</SelectItem>
-                  {categories.map((category) => (
-                    <SelectItem key={category.id} value={category.id}>
-                      {category.name}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-            
-            <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-              <div>
-                <Label htmlFor="cost_unit">Custo Unitário (R$) *</Label>
-                <Input
-                  id="cost_unit"
-                  type="number"
-                  step="0.01"
-                  value={formData.cost_unit}
-                  onChange={(e) => setFormData(prev => ({ ...prev, cost_unit: parseFloat(e.target.value) || 0 }))}
-                  required
-                />
-              </div>
-              
-              <div>
-                <Label htmlFor="packaging_cost">Custo da Embalagem (R$)</Label>
-                <Input
-                  id="packaging_cost"
-                  type="number"
-                  step="0.01"
-                  value={formData.packaging_cost}
-                  onChange={(e) => setFormData(prev => ({ ...prev, packaging_cost: parseFloat(e.target.value) || 0 }))}
-                />
-              </div>
-              
-              <div>
-                <Label htmlFor="tax_rate">Alíquota de Imposto (%)</Label>
-                <Input
-                  id="tax_rate"
-                  type="number"
-                  step="0.01"
-                  value={formData.tax_rate}
-                  onChange={(e) => setFormData(prev => ({ ...prev, tax_rate: parseFloat(e.target.value) || 0 }))}
-                />
-              </div>
-            </div>
-            
-            <div className="flex gap-2">
-              <Button type="submit" disabled={createMutation.isPending || updateMutation.isPending}>
-                {editingId ? "Atualizar" : "Criar"}
-              </Button>
-              {editingId && (
-                <Button type="button" variant="outline" onClick={resetForm}>
-                  Cancelar
-                </Button>
-              )}
-            </div>
-          </form>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader>
-          <CardTitle>Produtos Cadastrados</CardTitle>
-        </CardHeader>
-        <CardContent>
-          {isLoading ? (
-            <p>Carregando...</p>
-          ) : (
-            <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>Nome</TableHead>
-                  <TableHead>SKU</TableHead>
-                  <TableHead>Categoria</TableHead>
-                  <TableHead>Custo Unit.</TableHead>
-                  <TableHead>Ações</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {products.map((product) => (
-                  <TableRow key={product.id}>
-                    <TableCell className="font-medium">{product.name}</TableCell>
-                    <TableCell>{product.sku}</TableCell>
-                    <TableCell>{product.categories?.name || "Sem categoria"}</TableCell>
-                    <TableCell>{formatarMoeda(product.cost_unit || 0)}</TableCell>
-                    <TableCell>
-                      <div className="flex gap-2">
-                        <Button
-                          size="sm"
-                          variant="outline"
-                          onClick={() => handleEdit(product)}
-                        >
-                          <Edit className="h-4 w-4" />
-                        </Button>
-                        <Button
-                          size="sm"
-                          variant="destructive"
-                          onClick={() => deleteMutation.mutate(product.id)}
-                          disabled={deleteMutation.isPending}
-                        >
-                          <Trash2 className="h-4 w-4" />
-                        </Button>
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                ))}
-              </TableBody>
-            </Table>
-          )}
-        </CardContent>
-      </Card>
-    </div>
-  );
-};
